requirements/production.txt
-r base.txt

django-prometheus==2.4.1
django-redis==5.4.0
daphne==4.2.1
django-cachalot==2.8.0
channels-redis>=4.2,<5.0
psycopg[binary]>=3.2,<4.0
meilisearch>=0.34,<1.0
prometheus-client==0.21.1
python-json-logger==2.0.7

celery>=5.6.2,<6.0  # Python >=3.12 compatibility (Celery 5.6.2+)
